fix: ios callkit related audio inconsistencies (#1982)
### 💡 Overview

There is a niche audio bug described below:

1. join a call with callkit. callmanager.start()
2. leave the call. callmanager.stop()
3. join the same call without callkit. callmanager.start()

now the audio configuration for webrtc is not set. Mic cannot be
activated.

This is because callkit usage activates the audio session for webrtc.
But if the next call doesnt use it, the SDK does not take care that it
activates again. This is fixed in the PR

### 📝 Implementation notes

We explicitly set the activation flag to false for webrtc in
callmanager.stop() for consistency now. Just this change will fix it.
But additionally on callmanager.start() we ensure we activate always
because an integrator may have some other library that changes the audio
session without stop(). failsafe.

🎫 Ticket: https://linear.app/stream/issue/RN-299