fix(noise cancellation): delay toggling until initialization is finished (#2014)
### 💡 Overview

This PR fixes a race condition where setEnabled could run before the
Krisp SDK was fully initialized. In that use-case the filter node exists
but the promise is not yet resolved which leads to potential
`WASM_OR_WORKER_NOT_READY` errors.
With these changes toggle actions now wait for the initialization
promise to resolve before executing.


🎫 Ticket:
https://linear.app/stream/issue/REACT-667/krisp-delay-toggling-until-initialization-is-finished