fix: added call state update for handling case when call.ring event as not triggered (#2035)
### 💡 Overview

The case: 
when creating a ring type call with `getOrCreate` method, ws
`call.created` event triggers first (for some reason `call.ring` is not
detected) –> it creates a call instance with `ringing` property set to
false and registers it in the state store. After that we handle response
in `getOrCreateMethod` which tries to register the call (at this point
with `ringing` set to true) which is skipped as the call was already
registered. So in this scenario `useCalls` doesn't return the ringing
call.

### 📝 Implementation notes
The fix adds register or update method, which will override call
properties inside `get` and `getOrCreate` methods, which fixes behavior
in the case described above.

🎫 Ticket:
https://linear.app/stream/issue/RN-315/create-ring-type-call-issue