fix: revert usage of useSyncExternalStore (#2043)
Reverts #1953. In some environments, `useSyncExternalStore` causes
misleading warnings that the result of `getSnapshot` should be cached.
Although this isn't correct, I understand the confusion this warning
creates. Until we find a way to get rid of it, I'm reverting to the old
implementation of `useObservableValue`, based on `useState + useEffect`
combo.

Fixes: #2034 #2006
Ref: #2008